break up qick.py into separate files under drivers/
