+++ qick_lib/qick/qick.py
@@ -2,1977 +2,20 @@
 The lower-level driver for the QICK library. Contains classes for interfacing with the SoC.
 """
 import os
-from pynq.overlay import Overlay, DefaultIP
-from pynq.buffer import allocate
-try:
-    import xrfclk
-    import xrfdc
-except:
-    pass
+from pynq.overlay import Overlay
+import xrfclk
+import xrfdc
 import numpy as np
 import time
 import queue
-import logging
+from . import bitfile_path, obtain
+from .ip import SocIp, QickMetadata
 from .parser import parse_to_bin
 from .streamer import DataStreamer
-from .qick_asm import QickConfig, QickProgram, obtain
-from .helpers import QickMetadata
-from . import bitfile_path
-
-
-class SocIp(DefaultIP):
-    """
-    SocIp class
-    """
-    REGISTERS = {}
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        #print("SocIp init", description)
-        super().__init__(description)
-        self.fullpath = description['fullpath']
-        self.type = description['type'].split(':')[-2]
-        #self.ip = description
-        self.logger = logging.getLogger(self.type)
-
-    def __setattr__(self, a, v):
-        """
-        Sets the arguments associated with a register
-
-        :param a: Register specified by an offset value
-        :type a: int
-        :param v: value to be written
-        :type v: int
-        """
-        try:
-            index = self.REGISTERS[a]
-            self.mmio.array[index] = np.uint32(obtain(v))
-        except KeyError:
-            super().__setattr__(a, v)
-
-    def __getattr__(self, a):
-        """
-        Gets the arguments associated with a register
-
-        :param a: register name
-        :type a: str
-        :return: Register arguments
-        :rtype: *args object
-        """
-        try:
-            index = self.REGISTERS[a]
-            return self.mmio.array[index]
-        except KeyError:
-            return super().__getattribute__(a)
-
-
-class AbsSignalGen(SocIp):
-    """
-    Abstract class which defines methods that are common to different signal generators.
-    """
-    # This signal generator has a waveform memory.
-    HAS_WAVEFORM = False
-    # This signal generator is controlled by the tProc.
-    HAS_TPROC = False
-    # The DAC channel has a mixer.
-    HAS_MIXER = False
-    # Interpolation factor relating the generator and DAC sampling freqs.
-    FS_INTERPOLATION = 1
-    # Waveform samples per fabric clock.
-    SAMPS_PER_CLK = 1
-    # Name of the input driven by the tProc (if applicable).
-    TPROC_PORT = 's1_axis'
-    # Name of the input driven by the waveform DMA (if applicable).
-    WAVEFORM_PORT = 's0_axis'
-    # Maximum waveform amplitude.
-    MAXV = 2**15-2
-    # Scale factor between MAXV and the default maximum amplitude (necessary to avoid overshoot).
-    MAXV_SCALE = 1.0
-
-    # Configure this driver with links to the other drivers, and the signal gen channel number.
-    def configure(self, ch, rf, fs, axi_dma=None, axis_switch=None):
-        # Channel number corresponding to entry in the QickConfig list of gens.
-        self.ch = ch
-
-        if self.HAS_WAVEFORM:
-            # dma
-            self.dma = axi_dma
-
-            # Switch
-            self.switch = axis_switch
-
-            # Define buffer.
-            self.buff = allocate(shape=self.MAX_LENGTH, dtype=np.int32)
-
-        # RF data converter
-        self.rf = rf
-
-        # DAC sampling frequency.
-        self.fs_dac = fs
-
-        # DDS sampling frequency.
-        self.fs_dds = fs/self.FS_INTERPOLATION
-
-    def configure_connections(self, soc):
-        self.soc = soc
-
-        if self.HAS_TPROC:
-            # what tProc output port drives this generator?
-            # we will eventually also use this to find out which tProc drives this gen, for multi-tProc firmwares
-            ((block, port),) = soc.metadata.trace_bus(self.fullpath, self.TPROC_PORT)
-            while True:
-                blocktype = soc.metadata.mod2type(block)
-                if blocktype in ["axis_tproc64x32_x8", "axis_tproc_v2"]: # we're done
-                    break
-                elif blocktype == "axis_clock_converter":
-                    ((block, port),) = soc.metadata.trace_bus(block, 'S_AXIS')
-                elif blocktype == "axis_cdcsync_v1":
-                    # port name is of the form 'm4_axis' - follow corresponding input 's4_axis'
-                    ((block, port),) = soc.metadata.trace_bus(block, "s"+port[1:])
-                elif blocktype == "sg_translator":
-                    ((block, port),) = soc.metadata.trace_bus(block, "s_tproc_axis")
-                else:
-                    raise RuntimeError("failed to trace tProc port for %s - ran into unrecognized IP block %s" % (self.fullpath, block))
-            # ask the tproc to translate this port name to a channel number
-            self.tproc_ch = getattr(soc, block).port2ch(port)
-
-        if self.HAS_WAVEFORM:
-            # what switch port drives this generator?
-            ((block, port),) = soc.metadata.trace_bus(self.fullpath, self.WAVEFORM_PORT)
-            # port names are of the form 'M01_AXIS'
-            self.switch_ch = int(port.split('_')[0][1:])
-
-        # what RFDC port does this generator drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm_axis')
-        # might need to jump through an axis_register_slice
-        while True:
-            blocktype = soc.metadata.mod2type(block)
-            if blocktype == "usp_rf_data_converter": # we're done
-                break
-            elif blocktype == "axis_register_slice":
-                ((block, port),) = soc.metadata.trace_bus(block, "M_AXIS")
-            elif blocktype == "axis_register_slice_nb":
-                ((block, port),) = soc.metadata.trace_bus(block, "m_axis")
-            else:
-                raise RuntimeError("failed to trace RFDC port for %s - ran into unrecognized IP block %s" % (self.fullpath, block))
-        # port names are of the form 's00_axis'
-        self.dac = port[1:3]
-
-        #print("%s: switch %d, tProc ch %d, DAC tile %s block %s"%(self.fullpath, self.switch_ch, self.tproc_ch, *self.dac))
-
-    # Load waveforms.
-    def load(self, xin, addr=0):
-        """
-        Load waveform into I,Q envelope
-
-        :param xin: array of (I, Q) values for pulse envelope
-        :type xin: int16 array
-        :param addr: starting address
-        :type addr: int
-        """
-        if not self.HAS_WAVEFORM:
-            raise NotImplementedError(
-                "This generator does not support waveforms.")
-
-        length = xin.shape[0]
-        assert xin.dtype==np.int16
-
-        # Check for max length.
-        if length+addr > self.MAX_LENGTH:
-            raise RuntimeError("%s: buffer length must be %d samples or less." %
-                  (self.__class__.__name__, self.MAX_LENGTH))
-
-        # Check for even transfer size.
-        #if length % 2 != 0:
-        #    raise RuntimeError("Buffer transfer length must be even number.")
-
-        # Route switch to channel.
-        self.switch.sel(mst=self.switch_ch)
-
-        #print(self.fullpath, xin.shape, addr, self.switch_ch)
-
-        # Pack the data into a single array; columns will be concatenated
-        # -> lower 16 bits: I value.
-        # -> higher 16 bits: Q value.
-        # Format and copy data.
-        np.copyto(self.buff[:length],
-                np.frombuffer(xin, dtype=np.int32))
-
-        ################
-        ### Load I/Q ###
-        ################
-        # Enable writes.
-        self._wr_enable(addr)
-
-        # DMA data.
-        self.dma.sendchannel.transfer(self.buff, nbytes=int(length*4))
-        self.dma.sendchannel.wait()
-
-        # Disable writes.
-        self._wr_disable()
-
-    def _wr_enable(self, addr=0):
-        """
-           Enable WE reg
-        """
-        self.start_addr_reg = addr
-        self.we_reg = 1
-
-    def _wr_disable(self):
-        """
-           Disable WE reg
-        """
-        self.we_reg = 0
-
-    def set_nyquist(self, nqz):
-        self.rf.set_nyquist(self.dac, nqz)
-
-    def set_mixer_freq(self, f, ro_ch=None):
-        if not self.HAS_MIXER:
-            raise NotImplementedError("This channel does not have a mixer.")
-        if ro_ch is None:
-            rounded_f = f
-        else:
-            mixercfg = {}
-            mixercfg['fs'] = self.fs_dac
-            mixercfg['b_dds'] = 48
-            fstep = self.soc.calc_fstep(mixercfg, self.soc['readouts'][ro_ch])
-            rounded_f = round(f/fstep)*fstep
-        self.rf.set_mixer_freq(self.dac, rounded_f)
-
-    def get_mixer_freq(self):
-        if not self.HAS_MIXER:
-            raise NotImplementedError("This channel does not have a mixer.")
-        return self.rf.get_mixer_freq(self.dac)
-
-
-class AxisSignalGen(AbsSignalGen):
-    """
-    AxisSignalGen class
-    Supports AxisSignalGen V4+V5+V6, since they have the same software interface (ignoring registers that are not used)
-
-    AXIS Signal Generator Registers.
-    START_ADDR_REG
-
-    WE_REG
-    * 0 : disable writes.
-    * 1 : enable writes.
-    """
-    bindto = ['user.org:user:axis_signal_gen_v4:1.0',
-              'user.org:user:axis_signal_gen_v5:1.0',
-              'user.org:user:axis_signal_gen_v6:1.0']
-    REGISTERS = {'start_addr_reg': 0, 'we_reg': 1, 'rndq_reg': 2}
-    HAS_TPROC = True
-    HAS_WAVEFORM = True
-    SAMPS_PER_CLK = 16
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Default registers.
-        self.start_addr_reg = 0
-        self.we_reg = 0
-        self.rndq_reg = 10
-
-        # Generics
-        self.N = int(description['parameters']['N'])
-        self.NDDS = int(description['parameters']['N_DDS'])
-
-        # Maximum number of samples
-        self.MAX_LENGTH = 2**self.N*self.NDDS
-
-        # Frequency resolution
-        self.B_DDS = 32
-
-    def rndq(self, sel_):
-        """
-           TODO: remove this function. This functionality was removed from IP block.
-        """
-        self.rndq_reg = sel_
-
-class AxisSgInt4V1(AbsSignalGen):
-    """
-    AxisSgInt4V1
-
-    The default max amplitude for this generator is 0.9 times the maximum of int16.
-    This is necessary to prevent interpolation overshoot:
-    the output of the interpolation filter may exceed the max value of the input points.
-    (https://blogs.keysight.com/blogs/tech/rfmw.entry.html/2019/05/07/confronting_measurem-IBRp.html)
-    The result of overshoot is integer overflow in the filter output and big negative spikes.
-    If the input to the filter is a square pulse, the rising edge of the output overshoots by 10%.
-    Therefore, scaling envelopes by 90% seems safe.
-
-    AXIS Signal Generator with envelope x4 interpolation V1 Registers.
-    START_ADDR_REG
-
-    WE_REG
-    * 0 : disable writes.
-    * 1 : enable writes.
-    """
-    bindto = ['user.org:user:axis_sg_int4_v1:1.0']
-    REGISTERS = {'start_addr_reg': 0, 'we_reg': 1}
-    HAS_TPROC = True
-    HAS_WAVEFORM = True
-    HAS_MIXER = True
-    FS_INTERPOLATION = 4
-    MAXV_SCALE = 0.9
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Default registers.
-        self.start_addr_reg = 0
-        self.we_reg = 0
-
-        # Generics
-        self.N = int(description['parameters']['N'])
-        self.NDDS = 4  # Fixed by design, not accesible.
-
-        # Frequency resolution
-        self.B_DDS = 16
-
-        # Maximum number of samples
-        # Table is interpolated. Length is given only by parameter N.
-        self.MAX_LENGTH = 2**self.N
-
-
-class AxisSgMux4V1(AbsSignalGen):
-    """
-    AxisSgMux4V1
-
-    AXIS Signal Generator with 4 muxed outputs V1 registers.
-
-    PINC0_REG : frequency of tone 0.
-    PINC1_REG : frequency of tone 1.
-    PINC2_REG : frequency of tone 2.
-    PINC3_REG : frequency of tone 3.
-
-    WE_REG
-    * 0 : disable writes.
-    * 1 : enable writes.
-    """
-    bindto = ['user.org:user:axis_sg_mux4_v1:1.0']
-    REGISTERS = {'pinc0_reg': 0,
-            'pinc1_reg': 1,
-            'pinc2_reg': 2,
-            'pinc3_reg': 3,
-            'we_reg': 4}
-
-    HAS_TPROC = True
-    HAS_MIXER = True
-    FS_INTERPOLATION = 4
-    TPROC_PORT = 's_axis'
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Generics
-        self.NDDS = int(description['parameters']['N_DDS'])
-
-        # Frequency resolution
-        self.B_DDS = 16
-
-        # dummy values, since this doesn't have a waveform memory.
-        self.switch_ch = -1
-        self.MAX_LENGTH = 0
-
-        # Default registers.
-        self.pinc0_reg = 0
-        self.pinc1_reg = 0
-        self.pinc2_reg = 0
-        self.pinc3_reg = 0
-
-        self.update()
-
-    def update(self):
-        """
-        Update register values
-        """
-        self.we_reg = 1
-        self.we_reg = 0
-
-    def set_freq(self, f, out=0, ro_ch=0):
-        """
-        Set frequency register
-
-        :param f: frequency in MHz
-        :type f: float
-        :param out: muxed channel to configure
-        :type out: int
-        :param ro_ch: ADC channel (use None if you don't want to round to a valid ADC frequency)
-        :type ro_ch: int
-        """
-        # Sanity check.
-        k_i = np.int64(self.soc.freq2reg(f, gen_ch=self.ch, ro_ch=ro_ch))
-        self.set_freq_int(k_i, out)
-
-    def set_freq_int(self, k_i, out=0):
-        if out not in [0,1,2,3]:
-            raise IndexError("Invalid output index for mux.")
-        setattr(self, "pinc%d_reg" % (out), np.uint16(k_i))
-
-        # Register update.
-        self.update()
-
-    def get_freq(self, out=0):
-        return getattr(self, "pinc%d_reg" % (out)) * self.fs_dds / (2**self.B_DDS)
-
-class AxisSgMux4V2(AbsSignalGen):
-    """
-    AxisSgMux4V2
-
-    AXIS Signal Generator with 4 muxed outputs V2 registers.
-
-    PINC0_REG : frequency of tone 0.
-    PINC1_REG : frequency of tone 1.
-    PINC2_REG : frequency of tone 2.
-    PINC3_REG : frequency of tone 3.
-    GAIN0_REG : gain of tone 0.
-    GAIN1_REG : gain of tone 1.
-    GAIN2_REG : gain of tone 2.
-    GAIN3_REG : gain of tone 3.
-
-    WE_REG
-    * 0 : disable writes.
-    * 1 : enable writes.
-    """
-    bindto = ['user.org:user:axis_sg_mux4_v2:1.0']
-    REGISTERS = {'pinc0_reg':0,
-                 'pinc1_reg':1,
-                 'pinc2_reg':2,
-                 'pinc3_reg':3,
-                 'gain0_reg':4,
-                 'gain1_reg':5,
-                 'gain2_reg':6,
-                 'gain3_reg':7,
-                 'we_reg':8}
-
-    HAS_TPROC = True
-    HAS_MIXER = True
-    FS_INTERPOLATION = 4
-    TPROC_PORT = 's_axis'
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Generics
-        self.NDDS = int(description['parameters']['N_DDS'])
-
-        # Frequency resolution
-        self.B_DDS = 32
-
-        # dummy values, since this doesn't have a waveform memory.
-        self.switch_ch = -1
-        self.MAX_LENGTH = 0
-
-        # Default registers.
-        self.pinc0_reg=0
-        self.pinc1_reg=0
-        self.pinc2_reg=0
-        self.pinc3_reg=0
-        self.gain0_reg=self.MAXV
-        self.gain1_reg=self.MAXV
-        self.gain2_reg=self.MAXV
-        self.gain3_reg=self.MAXV
-
-        self.update()
-
-    def update(self):
-        """
-        Update register values
-        """
-        self.we_reg = 1
-        self.we_reg = 0
-
-    def set_freq(self, f, out, ro_ch=0):
-        """
-        Set frequency register
-
-        :param f: frequency in MHz
-        :type f: float
-        :param out: muxed channel to configure
-        :type out: int
-        :param ro_ch: ADC channel (use None if you don't want to round to a valid ADC frequency)
-        :type ro_ch: int
-        """
-        k_i = np.int64(self.soc.freq2reg(f, gen_ch=self.ch, ro_ch=ro_ch))
-        self.set_freq_int(k_i, out)
-
-    def set_freq_int(self, k_i, out):
-        if out not in range(4):
-            raise IndexError("Invalid output index for mux.")
-        setattr(self, "pinc%d_reg" % (out), np.uint32(k_i))
-
-        # Register update.
-        self.update()
-
-    def get_freq(self, out):
-        return getattr(self, "pinc%d_reg" % (out)) * self.fs_dds / (2**self.B_DDS)
-
-    def set_gain(self, g, out):
-        """
-        Set gain register
-
-        :param g: gain (in range -1 to 1)
-        :type g: float
-        :param out: muxed channel to configure
-        :type out: int
-        """
-        self.set_gain_int(np.round(g*self.MAXV), out)
-
-    def set_gain_int(self, g_i, out):
-        # Sanity checks.
-        if out not in range(4):
-            raise IndexError("Invalid output index for mux.")
-        if np.abs(g_i)>self.MAXV:
-            raise RuntimeError("Requested gain exceeds max limit.")
-        setattr(self, "gain%d_reg" % (out), np.int16(g_i))
-
-        # Register update.
-        self.update()
-
-class AxisConstantIQ(AbsSignalGen):
-    # AXIS Constant IQ registers:
-    # REAL_REG : 16-bit.
-    # IMAG_REG : 16-bit.
-    # WE_REG   : 1-bit. Update registers.
-    bindto = ['user.org:user:axis_constant_iq:1.0']
-    REGISTERS = {'real_reg': 0, 'imag_reg': 1, 'we_reg': 2}
-    HAS_MIXER = True
-
-    def __init__(self, description):
-        # Initialize ip
-        super().__init__(description)
-
-        # Default registers.
-        self.real_reg = self.MAXV
-        self.imag_reg = self.MAXV
-
-        # Register update.
-        self.update()
-
-    def update(self):
-        self.we_reg = 1
-        self.we_reg = 0
-
-    def set_iq(self, i=1, q=1):
-        # Set registers.
-        self.real_reg = np.int16(i*self.MAXV)
-        self.imag_reg = np.int16(q*self.MAXV)
-
-        # Register update.
-        self.update()
-
-
-class AxisReadoutV2(SocIp):
-    """
-    AxisReadoutV2 class
-
-    Registers.
-    FREQ_REG : 32-bit.
-
-    PHASE_REG : 32-bit.
-
-    NSAMP_REG : 16-bit.
-
-    OUTSEL_REG : 2-bit.
-    * 0 : product.
-    * 1 : dds.
-    * 2 : input (bypass).
-
-    MODE_REG : 1-bit.
-    * 0 : NSAMP.
-    * 1 : Periodic.
-
-    WE_REG : enable/disable to perform register update.
-
-    :param fs: sampling frequency in MHz
-    :type fs: float
-    """
-    bindto = ['user.org:user:axis_readout_v2:1.0']
-    REGISTERS = {'freq_reg': 0, 'phase_reg': 1, 'nsamp_reg': 2,
-                 'outsel_reg': 3, 'mode_reg': 4, 'we_reg': 5}
-
-    # Bits of DDS.
-    B_DDS = 32
-
-    # this readout is not controlled by the tProc.
-    tproc_ch = None
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Default registers.
-        self.freq_reg = 0
-        self.phase_reg = 0
-        self.nsamp_reg = 10
-        self.outsel_reg = 0
-        self.mode_reg = 1
-
-        # Register update.
-        self.update()
-
-    # Configure this driver with the sampling frequency.
-    def configure(self, fs):
-        # Sampling frequency.
-        self.fs = fs
-
-    def configure_connections(self, soc):
-        self.soc = soc
-
-        # what RFDC port drives this readout?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 's_axis')
-        # might need to jump through an axis_register_slice
-        while soc.metadata.mod2type(block) == "axis_register_slice":
-            ((block, port),) = soc.metadata.trace_bus(block, 'S_AXIS')
-        # port names are of the form 'm02_axis' where the block number is always even
-        iTile, iBlock = [int(x) for x in port[1:3]]
-        if soc.hs_adc:
-            iBlock //= 2
-        self.adc = "%d%d" % (iTile, iBlock)
-
-        # what buffer does this readout drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm1_axis')
-        self.buffer = getattr(soc, block)
-
-        #print("%s: ADC tile %s block %s, buffer %s"%(self.fullpath, *self.adc, self.buffer.fullpath))
-
-    def initialize(self):
-        """
-        Does nothing.
-        """
-        pass
-
-    def update(self):
-        """
-        Update register values
-        """
-        self.we_reg = 1
-        self.we_reg = 0
-
-    def set_out(self, sel="product"):
-        """
-        Select readout signal output
-
-        :param sel: select mux control
-        :type sel: int
-        """
-        self.outsel_reg = {"product": 0, "dds": 1, "input": 2}[sel]
-
-        # Register update.
-        self.update()
-
-    def set_freq(self, f, gen_ch=0):
-        """
-        Set frequency register
-
-        :param f: frequency in MHz (before adding any DAC mixer frequency)
-        :type f: float
-        :param gen_ch: DAC channel (use None if you don't want to round to a valid DAC frequency)
-        :type gen_ch: int
-        """
-        thiscfg = {}
-        thiscfg['fs'] = self.fs
-        thiscfg['b_dds'] = self.B_DDS
-        # calculate the exact frequency we expect to see
-        ro_freq = f
-        if gen_ch is not None: # calculate the frequency that will be applied to the generator
-            ro_freq = self.soc.roundfreq(f, self.soc['gens'][gen_ch], thiscfg)
-        if gen_ch is not None and self.soc.gens[gen_ch].HAS_MIXER:
-            ro_freq += self.soc.gens[gen_ch].get_mixer_freq()
-        ro_freq = ro_freq % self.fs
-        # we can calculate the register value without further referencing the gen_ch
-        f_int = self.soc.freq2int(ro_freq, thiscfg)
-        self.set_freq_int(f_int)
-
-    def set_freq_int(self, f_int):
-        """
-        Set frequency register (integer version)
-
-        :param f_int: frequency value register
-        :type f_int: int
-        """
-        self.freq_reg = np.int64(f_int)
-
-        # Register update.
-        self.update()
-
-    def get_freq(self):
-        return self.freq_reg * self.fs / (2**self.B_DDS)
-
-class AxisPFBReadoutV2(SocIp):
-    """
-    AxisPFBReadoutV2 class.
-
-    This readout block contains a polyphase filter bank with 8 channels.
-    Channel i mixes the input signal down by a fixed frequency f = i * fs/16,
-    then by a programmable DDS with a range of +/- fs/16.
-
-    The PFB channels can be freely mapped to the 4 outputs of the readout block.
-
-    Registers.
-    FREQ[0-7]_REG : 32-bit frequency of each channel.
-
-    OUTSEL_REG : 2-bit.
-    * 0 : product.
-    * 1 : input (bypass).
-    * 2 : dds.
-
-    CH[0-3]SEL_REG : 3-bit ID mapping an output channel to an input.
-    """
-    bindto = ['user.org:user:axis_pfb_readout_v2:1.0']
-    REGISTERS = {'freq0_reg': 0,
-            'freq1_reg': 1,
-            'freq2_reg': 2,
-            'freq3_reg': 3,
-            'freq4_reg': 4,
-            'freq5_reg': 5,
-            'freq6_reg': 6,
-            'freq7_reg': 7,
-            'outsel_reg': 8,
-            'ch0sel_reg': 9,
-            'ch1sel_reg': 10,
-            'ch2sel_reg': 11,
-            'ch3sel_reg': 12,
-            }
-
-    # Bits of DDS. 
-    # The channelizer DDS range is 1/8 of the sampling frequency, which effectively adds 3 bits of resolution.
-    B_DDS = 35
-
-    # index of the PFB channel that is centered around DC.
-    CH_OFFSET = 4
-
-    # this readout is not controlled by the tProc.
-    tproc_ch = None
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-        self.initialize()
-
-    # Configure this driver with the sampling frequency.
-    def configure(self, fs):
-        # Sampling frequency.
-        self.fs = fs
-
-    def configure_connections(self, soc):
-        self.soc = soc
-
-        # what RFDC port drives this readout?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 's_axis')
-        # might need to jump through an axis_register_slice
-        while soc.metadata.mod2type(block) == "axis_register_slice":
-            ((block, port),) = soc.metadata.trace_bus(block, 'S_AXIS')
-        if soc.metadata.mod2type(block) == "axis_combiner":
-            ((block, port),) = soc.metadata.trace_bus(block, 'S00_AXIS')
-
-        # port names are of the form 'm02_axis' where the block number is always even
-        iTile, iBlock = [int(x) for x in port[1:3]]
-        if soc.hs_adc:
-            iBlock //= 2
-        self.adc = "%d%d" % (iTile, iBlock)
-
-        # what buffers does this readout drive?
-        self.buffers=[]
-        for iBuf in range(4):
-            ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm%d_axis'%(iBuf))
-            self.buffers.append(getattr(soc, block))
-
-        #print("%s: ADC tile %s block %s, buffers[0] %s"%(self.fullpath, *self.adc, self.buffers[0].fullpath))
-
-    def initialize(self):
-        """
-        Set up local variables to track definitions of frequencies or readout modes.
-        """
-        self.ch_freqs = {}
-        self.sel = None
-        self.out_chs = {}
-
-    def set_out(self, sel="product"):
-        """
-        Select readout signal output
-
-        :param sel: select mux control
-        :type sel: int
-        """
-        if self.sel is not None and sel != self.sel:
-            raise RuntimeError("trying to set output mode to %s, but mode was previously set to %s"%(sel, self.sel))
-        self.sel = sel
-        self.outsel_reg = {"product": 0, "input": 1, "dds": 2}[sel]
-
-    def set_freq(self, f, out_ch, gen_ch=0):
-        """
-        Select the best PFB channel for reading out the requested frequency.
-        Set that channel's frequency register, and wire that channel to the specified output of the PFB readout block.
-
-        :param f: frequency in MHz (before adding any DAC mixer frequency)
-        :type f: float
-        :param out_ch: output channel
-        :type out_ch: int
-        :param gen_ch: DAC channel (use None if you don't want to round to a valid DAC frequency)
-        :type gen_ch: int
-        """
-        thiscfg = {}
-        thiscfg['fs'] = self.fs
-        thiscfg['b_dds'] = self.B_DDS
-        # calculate the exact frequency we expect to see
-        ro_freq = f
-        if gen_ch is not None: # calculate the frequency that will be applied to the generator
-            ro_freq = self.soc.roundfreq(f, self.soc['gens'][gen_ch], thiscfg)
-        if gen_ch is not None and self.soc.gens[gen_ch].HAS_MIXER:
-            ro_freq += self.soc.gens[gen_ch].get_mixer_freq()
-
-        nqz = int(ro_freq // (self.fs/2)) + 1
-        if nqz % 2 == 0: # even Nyquist zone
-            ro_freq *= -1
-        # the PFB channels are separated by half the DDS range
-        # round() gives you the single best channel
-        # floor() and ceil() would give you the 2 best channels
-        # if you have two RO frequencies close together, you might need to force one of them onto a non-optimal channel
-        f_steps = int(np.round(ro_freq/(self.fs/16)))
-        f_dds = ro_freq - f_steps*(self.fs/16)
-        in_ch = (self.CH_OFFSET + f_steps) % 8
-
-        # we can calculate the register value without further referencing the gen_ch
-        freq_int = self.soc.freq2int(f_dds, thiscfg)
-        self.set_freq_int(freq_int, in_ch, out_ch)
-
-    def set_freq_int(self, f_int, in_ch, out_ch):
-        if in_ch in self.ch_freqs and f_int != self.ch_freqs[in_ch]:
-            # we are already using this PFB channel, and it's set to a different frequency
-            # now do a bunch of math to print an informative message
-            centerfreq = ((in_ch - self.CH_OFFSET) % 8) * (self.fs/16)
-            lofreq = centerfreq - self.fs/32
-            hifreq = centerfreq + self.fs/32
-            thiscfg = {}
-            thiscfg['fs'] = self.fs
-            thiscfg['b_dds'] = self.B_DDS
-            oldfreq = centerfreq + self.soc.int2freq(self.ch_freqs[in_ch], thiscfg)
-            newfreq = centerfreq + self.soc.int2freq(f_int, thiscfg)
-            raise RuntimeError("frequency collision: tried to set PFB output %d to %f MHz and output %d to %f MHz, but both map to the PFB channel that is optimal for [%f, %f] (all freqs expressed in first Nyquist zone)"%(out_ch, newfreq, self.out_chs[in_ch], oldfreq, lofreq, hifreq))
-        self.ch_freqs[in_ch] = f_int
-        self.out_chs[in_ch] = out_ch
-        # wire the selected PFB channel to the output
-        setattr(self, "ch%dsel_reg"%(out_ch), in_ch)
-        # set the PFB channel's DDS frequency
-        setattr(self, "freq%d_reg"%(in_ch), f_int)
-
-class AxisReadoutV3():
-    """tProc-controlled readout block.
-    This isn't a PYNQ driver, since the block has no registers for PYNQ control.
-    We still need this class to represent the block and its connectivity.
-    """
-    # Bits of DDS.
-    B_DDS = 32
-
-    def __init__(self, fullpath):
-        self.fullpath = fullpath
-        self.type = "axis_readout_v3"
-
-    # Configure this driver with the sampling frequency.
-    def configure(self, fs):
-        # Sampling frequency.
-        self.fs = fs
-
-    def configure_connections(self, soc):
-        self.soc = soc
-
-        # what tProc output port controls this readout?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 's0_axis')
-        while True:
-            blocktype = soc.metadata.mod2type(block)
-            if blocktype == "axis_tproc64x32_x8": # we're done
-                break
-            elif blocktype == "axis_clock_converter":
-                ((block, port),) = soc.metadata.trace_bus(block, 'S_AXIS')
-            elif blocktype == "axis_cdcsync_v1":
-                # port name is of the form 'm4_axis' - follow corresponding input 's4_axis'
-                ((block, port),) = soc.metadata.trace_bus(block, "s"+port[1:])
-            else:
-                raise RuntimeError("failed to trace tProc port for %s - ran into unrecognized IP block %s" % (self.fullpath, block))
-        # port names are of the form 'm2_axis_tdata'
-        # subtract 1 to get the output channel number (m0 goes to the DMA)
-        self.tproc_ch = int(port.split('_')[0][1:])-1
-
-        # what RFDC port drives this readout?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 's1_axis')
-        while True:
-            blocktype = soc.metadata.mod2type(block)
-            if blocktype == "usp_rf_data_converter": # we're done
-                break
-            elif blocktype == "axis_resampler_2x1_v1":
-                ((block, port),) = soc.metadata.trace_bus(block, 's_axis')
-            elif blocktype == "axis_register_slice":
-                ((block, port),) = soc.metadata.trace_bus(block, 'S_AXIS')
-            elif blocktype == "axis_clock_converter":
-                ((block, port),) = soc.metadata.trace_bus(block, 'S_AXIS')
-            else:
-                raise RuntimeError("failed to trace tProc port for %s - ran into unrecognized IP block %s" % (self.fullpath, block))
-
-        # port names are of the form 'm02_axis' where the block number is always even
-        iTile, iBlock = [int(x) for x in port[1:3]]
-        if soc.hs_adc:
-            iBlock //= 2
-        self.adc = "%d%d" % (iTile, iBlock)
-
-        # what buffer does this readout drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm_axis')
-        self.buffer = getattr(soc, block)
-
-        #print("%s: ADC tile %s block %s, buffer %s"%(self.fullpath, *self.adc, self.buffer.fullpath))
-
-class AxisAvgBuffer(SocIp):
-    """
-    AxisAvgBuffer class
-
-    Registers.
-    AVG_START_REG
-    * 0 : Averager Disabled.
-    * 1 : Averager Enabled (started by external trigger).
-
-    AVG_ADDR_REG : start address to write results.
-
-    AVG_LEN_REG : number of samples to be added.
-
-    AVG_DR_START_REG
-    * 0 : do not send any data.
-    * 1 : send data using m0_axis.
-
-    AVG_DR_ADDR_REG : start address to read data.
-
-    AVG_DR_LEN_REG : number of samples to be read.
-
-    BUF_START_REG
-    * 0 : Buffer Disabled.
-    * 1 : Buffer Enabled (started by external trigger).
-
-    BUF_ADDR_REG : start address to write results.
-
-    BUF_LEN_REG : number of samples to be buffered.
-
-    BUF_DR_START_REG
-    * 0 : do not send any data.
-    * 1 : send data using m1_axis.
-
-    BUF_DR_ADDR_REG : start address to read data.
-
-    BUF_DR_LEN_REG : number of samples to be read.
-
-    :param axi_dma_avg: dma block for average buffers
-    :type axi_dma_avg: str
-    :param switch_avg: switch block for average buffers
-    :type switch_avg: str
-    :param axi_dma_buf: dma block for raw buffers
-    :type axi_dma_buf: str
-    :param switch_buf: switch block for raw buffers
-    :type switch_buf: str
-    :param channel: readout channel selection
-    :type channel: int
-    """
-    bindto = ['user.org:user:axis_avg_buffer:1.0']
-    REGISTERS = {'avg_start_reg': 0,
-                 'avg_addr_reg': 1,
-                 'avg_len_reg': 2,
-                 'avg_dr_start_reg': 3,
-                 'avg_dr_addr_reg': 4,
-                 'avg_dr_len_reg': 5,
-                 'buf_start_reg': 6,
-                 'buf_addr_reg': 7,
-                 'buf_len_reg': 8,
-                 'buf_dr_start_reg': 9,
-                 'buf_dr_addr_reg': 10,
-                 'buf_dr_len_reg': 11}
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Default registers.
-        self.avg_start_reg = 0
-        self.avg_dr_start_reg = 0
-        self.buf_start_reg = 0
-        self.buf_dr_start_reg = 0
-
-        # Generics
-        self.B = int(description['parameters']['B'])
-        self.N_AVG = int(description['parameters']['N_AVG'])
-        self.N_BUF = int(description['parameters']['N_BUF'])
-
-        # Maximum number of samples
-        self.AVG_MAX_LENGTH = 2**self.N_AVG
-        self.BUF_MAX_LENGTH = 2**self.N_BUF
-
-        # Preallocate memory buffers for DMA transfers.
-        self.avg_buff = allocate(shape=self.AVG_MAX_LENGTH, dtype=np.int64)
-        self.buf_buff = allocate(shape=self.BUF_MAX_LENGTH, dtype=np.int32)
-
-    # Configure this driver with links to the other drivers.
-    def configure(self, axi_dma_avg, switch_avg, axi_dma_buf, switch_buf):
-        # DMAs.
-        self.dma_avg = axi_dma_avg
-        self.dma_buf = axi_dma_buf
-
-        # Switches.
-        self.switch_avg = switch_avg
-        self.switch_buf = switch_buf
-
-    def configure_connections(self, soc):
-        # which readout drives this buffer?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 's_axis')
-        blocktype = soc.metadata.mod2type(block)
-        if blocktype == "axis_readout_v3":
-            # the V3 readout block has no registers, so it doesn't get a PYNQ driver
-            # so we initialize it here
-            self.readout = AxisReadoutV3(block)
-            self.readout.configure_connections(soc)
-        else:
-            self.readout = getattr(soc, block)
-            if blocktype == "axis_pfb_readout_v2":
-                # port names are of the form 'm1_axis'
-                self.readoutport = int(port.split('_')[0][1:], 10)
-
-        # which switch_avg port does this buffer drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm0_axis')
-        # port names are of the form 'S01_AXIS'
-        switch_avg_ch = int(port.split('_')[0][1:], 10)
-
-        # which switch_buf port does this buffer drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm1_axis')
-        # port names are of the form 'S01_AXIS'
-        switch_buf_ch = int(port.split('_')[0][1:], 10)
-        if switch_avg_ch != switch_buf_ch:
-            raise RuntimeError(
-                "switch_avg and switch_buf port numbers do not match:", self.fullpath)
-        self.switch_ch = switch_avg_ch
-
-        # which tProc output bit triggers this buffer?
-        ((block, port),) = soc.metadata.trace_sig(self.fullpath, 'trigger')
-        # port names are of the form 'dout14'
-        self.trigger_bit = int(port[4:])
-
-        # which tProc input port does this buffer drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm2_axis')
-        # jump through an axis_clk_cnvrt
-        while soc.metadata.mod2type(block) == "axis_clock_converter":
-            ((block, port),) = soc.metadata.trace_bus(block, 'M_AXIS')
-        # port names are of the form 's1_axis'
-        # subtract 1 to get the channel number (s0 comes from the DMA)
-        if soc.metadata.mod2type(block) in ["axis_tproc64x32_x8", "axis_tproc_v2"]:
-            # ask the tproc to translate this port name to a channel number
-            self.tproc_ch = getattr(soc, block).port2ch(port)
-        else:
-            # this buffer doesn't feed back into the tProc
-            self.tproc_ch = -1
-
-        # print("%s: readout %s, switch %d, trigger %d, tProc port %d"%
-        # (self.fullpath, self.readout.fullpath, self.switch_ch, self.trigger_bit, self.tproc_ch))
-
-    def set_freq(self, f, gen_ch=0):
-        """
-        Set the downconversion frequency on the readout that drvies this buffer.
-
-        :param f: frequency in MHz (before adding any DAC mixer frequency)
-        :type f: float
-        :param gen_ch: DAC channel (use None if you don't want to round to a valid DAC frequency)
-        :type gen_ch: int
-        """
-        if isinstance(self.readout, AxisPFBReadoutV2):
-            self.readout.set_freq(f, self.readoutport, gen_ch=gen_ch)
-        else:
-            self.readout.set_freq(f, gen_ch=gen_ch)
-
-    def config(self, address=0, length=100):
-        """
-        Configure both average and raw buffers
-
-        :param addr: Start address of first capture
-        :type addr: int
-        :param length: window size
-        :type length: int
-        """
-        # Configure averaging and buffering to the same address and length.
-        self.config_avg(address=address, length=length)
-        self.config_buf(address=address, length=length)
-
-    def enable(self):
-        """
-        Enable both average and raw buffers
-        """
-        # Enable both averager and buffer.
-        self.enable_avg()
-        self.enable_buf()
-
-    def config_avg(self, address=0, length=100):
-        """
-        Configure average buffer data from average and buffering readout block
-
-        :param addr: Start address of first capture
-        :type addr: int
-        :param length: window size
-        :type length: int
-        """
-        # Disable averaging.
-        self.disable_avg()
-
-        # Set registers.
-        self.avg_addr_reg = address
-        self.avg_len_reg = length
-
-    def transfer_avg(self, address=0, length=100):
-        """
-        Transfer average buffer data from average and buffering readout block.
-
-        :param addr: starting reading address
-        :type addr: int
-        :param length: number of samples
-        :type length: int
-        :return: I,Q pairs
-        :rtype: list
-        """
-
-        if length % 2 != 0:
-            raise RuntimeError("Buffer transfer length must be even number.")
-        if length >= self.AVG_MAX_LENGTH:
-            raise RuntimeError("length=%d longer than %d" %
-                               (length, self.AVG_MAX_LENGTH))
-
-        # Route switch to channel.
-        self.switch_avg.sel(slv=self.switch_ch)
-
-        # Set averager data reader address and length.
-        self.avg_dr_addr_reg = address
-        self.avg_dr_len_reg = length
-
-        # Start send data mode.
-        self.avg_dr_start_reg = 1
-
-        # DMA data.
-        buff = self.avg_buff
-        # nbytes has to be a Python int (it gets passed to mmio.write, which requires int or bytes)
-        self.dma_avg.recvchannel.transfer(buff, nbytes=int(length*8))
-        self.dma_avg.recvchannel.wait()
-
-        # Stop send data mode.
-        self.avg_dr_start_reg = 0
-
-        if self.dma_avg.recvchannel.transferred != length*8:
-            raise RuntimeError("Requested %d samples but only got %d from DMA" % (
-                length, self.dma_avg.recvchannel.transferred//8))
-
-        # Format:
-        # -> lower 32 bits: I value.
-        # -> higher 32 bits: Q value.
-        data = np.frombuffer(buff[:length], dtype=np.int32).reshape((-1,2))
-
-        # data is a view into the data buffer, so copy it before returning
-
-        return data.copy()
-
-    def enable_avg(self):
-        """
-        Enable average buffer capture
-        """
-        self.avg_start_reg = 1
-
-    def disable_avg(self):
-        """
-        Disable average buffer capture
-        """
-        self.avg_start_reg = 0
-
-    def config_buf(self, address=0, length=100):
-        """
-        Configure raw buffer data from average and buffering readout block
-
-        :param addr: Start address of first capture
-        :type addr: int
-        :param length: window size
-        :type length: int
-        """
-        # Disable buffering.
-        self.disable_buf()
-
-        # Set registers.
-        self.buf_addr_reg = address
-        self.buf_len_reg = length
-
-    def transfer_buf(self, address=0, length=100):
-        """
-        Transfer raw buffer data from average and buffering readout block
-
-        :param addr: starting reading address
-        :type addr: int
-        :param length: number of samples
-        :type length: int
-        :return: I,Q pairs
-        :rtype: list
-        """
-
-        if length % 2 != 0:
-            raise RuntimeError("Buffer transfer length must be even number.")
-        if length >= self.BUF_MAX_LENGTH:
-            raise RuntimeError("length=%d longer or equal to %d" %
-                               (length, self.BUF_MAX_LENGTH))
-
-        # Route switch to channel.
-        self.switch_buf.sel(slv=self.switch_ch)
-
-        # time.sleep(0.050)
-
-        # Set buffer data reader address and length.
-        self.buf_dr_addr_reg = address
-        self.buf_dr_len_reg = length
-
-        # Start send data mode.
-        self.buf_dr_start_reg = 1
-
-        # DMA data.
-        buff = self.buf_buff
-        # nbytes has to be a Python int (it gets passed to mmio.write, which requires int or bytes)
-        self.dma_buf.recvchannel.transfer(buff, nbytes=int(length*4))
-        self.dma_buf.recvchannel.wait()
-
-        if self.dma_buf.recvchannel.transferred != length*4:
-            raise RuntimeError("Requested %d samples but only got %d from DMA" % (
-                length, self.dma_buf.recvchannel.transferred//4))
-
-        # Stop send data mode.
-        self.buf_dr_start_reg = 0
-
-        # Format:
-        # -> lower 16 bits: I value.
-        # -> higher 16 bits: Q value.
-        data = np.frombuffer(buff[:length], dtype=np.int16).reshape((-1,2))
-
-        # data is a view into the data buffer, so copy it before returning
-        return data.copy()
-
-    def enable_buf(self):
-        """
-        Enable raw buffer capture
-        """
-        self.buf_start_reg = 1
-
-    def disable_buf(self):
-        """
-        Disable raw buffer capture
-        """
-        self.buf_start_reg = 0
-
-
-class MrBufferEt(SocIp):
-    # Registers.
-    # DW_CAPTURE_REG
-    # * 0 : Capture disabled.
-    # * 1 : Capture enabled (capture started by external trigger).
-    #
-    # DR_START_REG
-    # * 0 : don't send.
-    # * 1 : start sending data.
-    #
-    # DW_CAPTURE_REG needs to be de-asserted and asserted again to allow a new capture.
-    # DR_START_REG needs to be de-assereted and asserted again to allow a new transfer.
-    #
-    bindto = ['user.org:user:mr_buffer_et:1.0']
-    REGISTERS = {'dw_capture_reg': 0, 'dr_start_reg': 1}
-
-    def __init__(self, description):
-        # Init IP.
-        super().__init__(description)
-
-        # Default registers.
-        self.dw_capture_reg = 0
-        self.dr_start_reg = 0
-
-        # Generics
-        self.B = int(description['parameters']['B'])
-        self.N = int(description['parameters']['N'])
-        self.NM = int(description['parameters']['NM'])
-
-        # Maximum number of samples
-        self.MAX_LENGTH = 2**self.N * self.NM
-
-        # Preallocate memory buffers for DMA transfers.
-        #self.buff = allocate(shape=self.MAX_LENGTH, dtype=np.int32)
-        self.buff = allocate(shape=self.MAX_LENGTH, dtype=np.int16)
-
-    def config(self, dma, switch):
-        self.dma = dma
-        self.switch = switch
-
-    def route(self, ch):
-        # Route switch to channel.
-        self.switch.sel(slv=ch)
-
-    def transfer(self, buff=None):
-        if buff is None:
-            buff = self.buff
-        # Start send data mode.
-        self.dr_start_reg = 1
-
-        # DMA data.
-        self.dma.recvchannel.transfer(buff)
-        self.dma.recvchannel.wait()
-
-        # Stop send data mode.
-        self.dr_start_reg = 0
-
-        return buff
-
-    def enable(self):
-        self.dw_capture_reg = 1
-
-    def disable(self):
-        self.dw_capture_reg = 0
-
-
-class AxisTProc64x32_x8(SocIp):
-    """
-    AxisTProc64x32_x8 class
-
-    AXIS tProcessor registers:
-    START_SRC_REG
-    * 0 : internal start (using START_REG)
-    * 1 : external start (using "start" input)
-
-    Regardless of the START_SRC, the start logic triggers on a rising edge:
-    A low level arms the trigger (transitions from "end" to "init" state).
-    A high level fires the trigger (starts the program).
-    To stop a running program, see reset().
-
-    START_REG
-    * 0 : init
-    * 1 : start
-
-    MEM_MODE_REG
-    * 0 : AXIS Read (from memory to m0_axis)
-    * 1 : AXIS Write (from s0_axis to memory)
-
-    MEM_START_REG
-    * 0 : Stop.
-    * 1 : Execute operation (AXIS)
-
-    MEM_ADDR_REG : starting memory address for AXIS read/write mode.
-
-    MEM_LEN_REG : number of samples to be transferred in AXIS read/write mode.
-
-    DMEM: The internal data memory is 2^DMEM_N samples, 32 bits each.
-    The memory can be accessed either single read/write from AXI interface. The lower 256 Bytes are reserved for registers.
-    The memory is then accessed in the upper section (beyond 256 bytes). Byte to sample conversion needs to be performed.
-    The other method is to DMA in and out. Here the access is direct, so no conversion is needed.
-    There is an arbiter to ensure data coherency and avoid blocking transactions.
-
-    :param mem: memory address
-    :type mem: int
-    :param axi_dma: axi_dma address
-    :type axi_dma: int
-    """
-    bindto = ['user.org:user:axis_tproc64x32_x8:1.0']
-    REGISTERS = {'start_src_reg': 0,
-                 'start_reg': 1,
-                 'mem_mode_reg': 2,
-                 'mem_start_reg': 3,
-                 'mem_addr_reg': 4,
-                 'mem_len_reg': 5}
-
-    # Number of 32-bit words in the lower address map (reserved for register access)
-    NREG = 64
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Default registers.
-        # start_src_reg = 0   : internal start.
-        # start_reg     = 0   : stopped.
-        # mem_mode_reg  = 0   : axis read.
-        # mem_start_reg = 0   : axis operation stopped.
-        # mem_addr_reg  = 0   : start address = 0.
-        # mem_len_reg   = 100 : default length.
-        self.start_src_reg = 0
-        self.start_reg = 0
-        self.mem_mode_reg = 0
-        self.mem_start_reg = 0
-        self.mem_addr_reg = 0
-        self.mem_len_reg = 100
-
-        # Generics.
-        # data memory address size (log2 of the number of 32-bit words)
-        self.DMEM_N = int(description['parameters']['DMEM_N'])
-        # program memory address size (log2 of the number of 64-bit words, though the actual memory is usually smaller)
-        self.PMEM_N = int(description['parameters']['PMEM_N'])
-
-    # Configure this driver with links to its memory and DMA.
-    # TODO: is this "mem" argument actually used? we are not setting it to anything sensible.
-    def configure(self, mem, axi_dma):
-        # Program memory.
-        self.mem = mem
-
-        # dma
-        self.dma = axi_dma
-
-    def configure_connections(self, soc):
-        self.output_pins = []
-        self.start_pin = None
-        try:
-            ((port),) = soc.metadata.trace_sig(self.fullpath, 'start')
-            # check if the start pin is driven by a port of the top-level design
-            if len(port)==1:
-                self.start_pin = port[0]
-        except:
-            pass
-        # search for the trigger port
-        for i in range(8):
-            # what block does this output drive?
-            # add 1, because output 0 goes to the DMA
-            try:
-                ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm%d_axis' % (i+1))
-            except: # skip disconnected tProc outputs
-                continue
-            if soc.metadata.mod2type(block) == "axis_set_reg":
-                self.trig_output = i
-                ((block, port),) = soc.metadata.trace_sig(block, 'dout')
-                for iPin in range(16):
-                    try:
-                        ports = soc.metadata.trace_sig(block, "dout%d"%(iPin))
-                        if len(ports)==1 and len(ports[0])==1:
-                            # it's an FPGA pin, save it
-                            pinname = ports[0][0]
-                            self.output_pins.append((iPin, pinname))
-                    except KeyError:
-                        pass
-
-    def port2ch(self, portname):
-        """
-        Translate a port name to a channel number.
-        Used in connection mapping.
-        """
-        # port names are of the form 'm2_axis' (for outputs) and 's2_axis (for inputs)
-        # subtract 1 to get the output channel number (s0/m0 goes to the DMA)
-        return int(portname.split('_')[0][1:])-1
-
-    def start(self):
-        """
-        Start tProc from register.
-        This has no effect if the tProc is not in init or end state,
-        or if the start source is set to "external."
-        """
-        self.start_reg = 0
-        self.start_reg = 1
-
-    def reset(self):
-        """
-        Force the tProc to stop by filling the program memory with "end" instructions.
-        For speed, we hard-code the "end" instruction and write directly to the program memory.
-        This typically takes about 1 ms.
-        """
-        # we only write the high half of each program word, the low half doesn't matter
-        np.copyto(self.mem.mmio.array[1::2],np.uint32(0x3F000000))
-
-    def single_read(self, addr):
-        """
-        Reads one sample of tProc data memory using AXI access
-
-        :param addr: reading address
-        :type addr: int
-        :return: requested value
-        :rtype: int
-        """
-        # Read data.
-        # Address should be translated to upper map.
-        return self.mmio.array[addr + self.NREG]
-
-    def single_write(self, addr=0, data=0):
-        """
-        Writes one sample of tProc data memory using AXI access
-
-        :param addr: writing address
-        :type addr: int
-        :param data: value to be written
-        :type data: int
-        """
-        # Write data.
-        # Address should be translated to upper map.
-        self.mmio.array[addr + self.NREG] = np.uint32(data)
-
-    def load_dmem(self, buff_in, addr=0):
-        """
-        Writes tProc data memory using DMA
-
-        :param buff_in: Input buffer
-        :type buff_in: int
-        :param addr: Starting destination address
-        :type addr: int
-        """
-        # Length.
-        length = len(buff_in)
-
-        # Configure dmem arbiter.
-        self.mem_mode_reg = 1
-        self.mem_addr_reg = addr
-        self.mem_len_reg = length
-
-        # Define buffer.
-        self.buff = allocate(shape=length, dtype=np.int32)
-
-        # Copy buffer.
-        np.copyto(self.buff, buff_in)
-
-        # Start operation on block.
-        self.mem_start_reg = 1
-
-        # DMA data.
-        self.dma.sendchannel.transfer(self.buff)
-        self.dma.sendchannel.wait()
-
-        # Set block back to single mode.
-        self.mem_start_reg = 0
-
-    def read_dmem(self, addr=0, length=100):
-        """
-        Reads tProc data memory using DMA
-
-        :param addr: Starting address
-        :type addr: int
-        :param length: Number of samples
-        :type length: int
-        :return: List of memory data
-        :rtype: list
-        """
-        # Configure dmem arbiter.
-        self.mem_mode_reg = 0
-        self.mem_addr_reg = addr
-        self.mem_len_reg = length
-
-        # Define buffer.
-        buff = allocate(shape=length, dtype=np.int32)
-
-        # Start operation on block.
-        self.mem_start_reg = 1
-
-        # DMA data.
-        self.dma.recvchannel.transfer(buff)
-        self.dma.recvchannel.wait()
-
-        # Set block back to single mode.
-        self.mem_start_reg = 0
-
-        return buff
-
-    class AxisTProc_v2(SocIp):
-        """
-        AxisTProc_v2 class
-     
-        AXIS tProcessor registers:
-        TPROC_CTRL       Write / Read 32-Bits
-        RAND             Read Only    32-Bits
-        TPROC_CFG        Write / Read  2-Bits
-        MEM_ADDR         Write / Read 16-Bits
-        MEM_LEN          Write / Read 16-Bits
-        MEM_DT_I         Write / Read 32-Bits
-        TPROC_EXT_DT1_I  Write / Read 32-Bits
-        TPROC_EXT_DT2_I  Write / Read 32-Bits
-        PORT_LSW         Read Only    32-Bits
-        PORT_MSW         Read Only    32-Bits
-        TIME_USR         Read Only    32-Bits
-        TPROC_EXT_DT1_O  Read Only    32-Bits
-        TPROC_EXT_DT2_O  Read Only    32-Bits
-        MEM_DT_O         Read Only    32-Bits
-        TPROC_STATUS     Read Only    32-Bits
-        TPROC_DEBUG      Read Only    32-Bits
-     
-        TPROC_CTRL[0] - Reset       : Reset the tProc
-        TPROC_CTRL[1] - Stop        : Stop the tProc
-        TPROC_CTRL[2] - Pause       : Pause the tProc(Time continue Running)
-        TPROC_CTRL[3] - Freeze      : Freeze Time (tProc Runs, but time stops)
-        TPROC_CTRL[4] - Play        : Starts / Continue running the tProc
-        TPROC_CTRL[10] - COND_set   : Set External Condition Flag from
-        TPROC_CTRL[11] - COND_clear : Clears External Condition Flag from
-        TPROC_CFG[0] - START_REG
-        * 0 : init
-        * 1 : Start
-        TPROC_CFG[1] - OPERATION_REG
-        * 0 : Read
-        * 1 : Write
-        TPROC_CFG[3:2] - MEM_BANK_REG 
-        * 0 : None Selected
-        * 1 : Program memory 
-        * 2 : Data Memory
-        * 3 : WaveParam Memory
-        TPROC_CFG[4] - SOURCE_REG
-        * 0 : AXIS Operation     (Using MEM_ADDR, MEM_LEN, (s0_axis / m0_axis) )
-        * 1 : REGISTER Operation (Using MEM_ADDR, (MEM_DT_I, MEM_DT_O) )
-     
-        MEM_ADDR : starting memory address for AXIS read/write mode.
-     
-        MEM_LEN : number of samples to be transferred in AXIS read/write mode.
-     
-        :param mem: memory address
-        :type mem: int
-        :param axi_dma: axi_dma address
-        :type axi_dma: int
-        """
-        bindto = ['Fermi:user:axis_tproc_v2:2.0']
-     
-        REGISTERS = {
-           'tproc_ctrl':0, 'rand':1, 'tproc_cfg':2,
-           'mem_addr'  :3, 'mem_len'  :4, 'mem_dt_i':5, 'mem_dt_o':6,
-           'tproc_ext_dt1_i':7 , 'tproc_ext_dt2_i':8,
-           'port_lsw':9, 'port_msw':10, 'time_usr':11,
-           'tproc_ext_dt1_o':12, 'tproc_ext_dt2_o':13,
-           'tproc_status':14, 'tproc_debug':15,}
-     
-        def __init__(self, description):
-            """
-            Constructor method
-            """
-            super().__init__(description)
-
-            # Parameters
-            self.PMEM_SIZE = pow( 2, int(description['parameters']['PMEM_AW']) )
-            self.DMEM_SIZE = pow( 2, int(description['parameters']['DMEM_AW']) )
-            self.WMEM_SIZE = pow( 2, int(description['parameters']['WMEM_AW']) )
-            self.DREG_QTY  = pow( 2, int(description['parameters']['REG_AW'])  )
-            self.IN_PORT_QTY   = int(description['parameters']['IN_PORT_QTY'])
-            self.OUT_DPORT_QTY = int(description['parameters']['OUT_DPORT_QTY'])
-            self.OUT_WPORT_QTY = int(description['parameters']['OUT_WPORT_QTY'])
-            self.LFSR      = int(description['parameters']['LFSR'])
-            self.DIVIDER   = int(description['parameters']['DIVIDER'])
-            self.ARITH     = int(description['parameters']['ARITH'])
-            self.TIME_CMP  = int(description['parameters']['TIME_CMP'])
-            self.TIME_READ = int(description['parameters']['TIME_READ'])
-            
-            # Initial Values 
-            self.tproc_ctrl = 0
-            self.tproc_cfg  = 0
-            self.mem_addr   = 0
-            self.mem_len    = 100
-            self.mem_dt_i   = 0
-            self.tproc_ext_dt1_i = 0
-            self.tproc_ext_dt2_i = 0
-            
-            #COmpatible with previous Version
-            self.DMEM_N = int(description['parameters']['DMEM_AW']) 
-
-      
-        # Configure this driver with links to its memory and DMA.
-        def configure(self, mem, axi_dma):
-            # Program memory.
-            self.mem = mem
-            # dma
-            self.dma = axi_dma
-     
-        def configure_connections(self, soc):
-            self.output_pins = []
-            self.start_pin = None
-            try:
-                ((port),) = soc.metadata.trace_sig(self.fullpath, 'start')
-                self.start_pin = port[0]
-            except:
-                pass
-           # search for the trigger port
-            for i in range(4):
-                
-                # what block does this output drive?
-                # add 1, because output 0 goes to the DMA
-                try:
-                    ((block, port),) = soc.metadata.trace_sig(self.fullpath, 'port_%d_dt_o' % (i))
-                except: # skip disconnected tProc outputs
-                    continue
-                if soc.metadata.mod2type(block).startswith("vect2bits"):
-                    self.trig_output = i
-                    for iPin in range(16):
-                        try:
-                            #print(iPin, trace_net(sigparser, block, "dout%d"%(iPin)))
-                            ports = soc.metadata.trace_sig(block, "dout%d"%(iPin))
-                            if len(ports)==1 and len(ports[0])==1:
-                                # it's an FPGA pin, save it
-                                pinname = ports[0][0]
-                                self.output_pins.append((iPin, pinname))
-                        except KeyError:
-                            pass
-
-        def port2ch(self, portname):
-            """
-            Translate a port name to a channel number.
-            Used in connection mapping.
-            """
-            # port names are of the form 'm2_axis' (for outputs) and 's2_axis (for inputs)
-            return int(portname.split('_')[0][1:])
-                        
-        def reset(self):
-            self.logger.info('RESET')
-            self.tproc_ctrl      = 1
-        def stop(self):
-            self.logger.info('STOP')
-            self.tproc_ctrl      = 2
-        def pause(self):
-            self.logger.info('PAUSE')
-            self.tproc_ctrl      = 4
-        def freeze(self):
-            self.logger.info('FREEZE')
-            self.tproc_ctrl      = 8
-        def run(self):
-            self.logger.info('RUN')
-            self.tproc_ctrl      = 16
-        def set_cond(self):
-            self.logger.info('RESET')
-            self.tproc_ctrl      = 1024
-        def clear_cond(self):
-            self.logger.info('RESET')
-            self.tproc_ctrl      = 2048
-            
-        def info(self):
-            print(self)
-            
-        def __str__(self):
-            lines = []
-            lines.append('---------------------------------------------')
-            lines.append(' TPROC V2 INFO ')
-            lines.append('---------------------------------------------')
-            for param in ["PMEM_SIZE", "DMEM_SIZE", "WMEM_SIZE", "DREG_QTY", "IN_PORT_QTY", "OUT_DPORT_QTY", "OUT_WPORT_QTY"]:
-                lines.append("%-14s: %d" % (param, getattr(self, param)))
-            lines.append("\nPeripherals:")
-            for param in ["LFSR", "DIVIDER", "ARITH", "TIME_CMP", "TIME_READ"]:
-                lines.append("%-14s: %s" % (param, ["NO", "YES"][getattr(self, param)]))
-            return "\n".join(lines)
-
-        def single_read(self, addr):
-            """
-            Reads one sample of tProc data memory using AXI access
-           
-            :param addr: reading address
-            :type addr: int
-            :return: requested value
-            :rtype: int
-            """
-            # Read data.
-            # Address should be translated to upper map.
-            return self.mmio.array[addr + self.NREG]
-     
-        def single_write(self, addr=0, data=0):
-            """
-            Writes one sample of tProc data memory using AXI access
-            
-            :param addr: writing address
-            :type addr: int
-            :param data: value to be written
-            :type data: int
-            """
-            # Write data.
-            # Address should be translated to upper map.
-            self.mmio.array[addr + self.NREG] = np.uint32(data)
-     
-
-        def load_mem(self,mem_sel, buff_in, addr=0):
-            """
-            Writes tProc Selected memory using DMA
-            PARAMETERS> 
-              mem_sel   : Destination Memory ( int PMEM=1, DMEM=2, WMEM=3 )
-              buff_in   : Input buffer ( int )
-              addr      : Starting destination address ( int )
-            """
-            # Length.
-            length = len(buff_in)
-            # Configure Memory arbiter. (Write MEM)
-            self.mem_addr        = addr
-            self.mem_len         = length
-
-            # Define buffer.
-            # TODO: pre-allocate buffer
-            self.buff = allocate(shape=(length,8), dtype=np.int32)
-            # Copy buffer.
-            np.copyto(self.buff, buff_in)
-            print(self.buff)
-            #Start operation
-            if (mem_sel==1):       # WRITE PMEM
-                self.tproc_cfg       = 7
-            elif (mem_sel==2):     # WRITE DMEM
-                self.tproc_cfg       = 11
-            elif (mem_sel==3):     # WRITE WMEM
-                self.tproc_cfg       = 15
-            else:
-                raise RuntimeError('Destination Memeory error should be  PMEM=1, DMEM=2, WMEM=3 current Value : %d' % (mem_sel))
-
-            # DMA data.
-            self.dma.sendchannel.transfer(self.buff)
-            self.dma.sendchannel.wait()
-            
-            # End Operation
-            self.tproc_cfg       = 0
-
-        
-        
-        def read_mem(self,mem_sel, addr=0, length=100):
-            """
-            Read tProc Selected memory using DMA
-            PARAMETERS> 
-              mem_sel   : Destination Memory ( int PMEM=1, DMEM=2, WMEM=3 )
-              buff_in   : Input buffer ( int )
-              addr      : Starting destination address ( int )
-            """
-        # Configure Memory arbiter. (Read DMEM)
-            self.mem_addr        = addr
-            self.mem_len         = length
-
-            # Define buffer.
-            # TODO: pre-allocate buffer
-            buff_rd = allocate(shape=(length,8), dtype=np.int32)
-
-            #Start operation
-            if (mem_sel==1):       # READ PMEM
-                self.tproc_cfg       = 5
-            elif (mem_sel==2):     # READ DMEM
-                self.tproc_cfg       = 9
-            elif (mem_sel==3):     # READ WMEM
-                self.tproc_cfg       = 13
-            else:
-                raise RuntimeError('Source Memeory error should be PMEM=1, DMEM=2, WMEM=3 current Value : %d' % (mem_sel))
-
-            # DMA data.
-            self.dma.recvchannel.transfer(buff_rd)
-            self.dma.recvchannel.wait()
-            
-            # End Operation
-            self.tproc_cfg       = 0      
-            return buff_rd
-        
-        def Load_PMEM(self, p_mem):
-            self.logger.info('Loading Program in PMEM')
-            # Length.
-            length = len(p_mem)
-            # Configure Memory arbiter.
-            self.mem_addr        = 0
-            self.mem_len         = length
-            # Define buffer.
-            self.buff = allocate(shape=(length,8), dtype=np.int32)
-            # Copy buffer.
-            np.copyto(self.buff, p_mem)
-            #Start operation
-            self.tproc_cfg       = 7
-            # DMA data.
-            self.logger.debug('P1')
-            self.dma.sendchannel.transfer(self.buff)
-            self.logger.debug('P2')
-            self.dma.sendchannel.wait()
-            self.logger.debug('P3')
-            # End Operation
-            self.tproc_cfg       = 0
-            
-            #Read PROGRAM MEMORY
-            # Configure Memory arbiter.
-            self.mem_addr        = 0
-            self.mem_len         = length
-            self.tproc_cfg       = 5
-            # DMA data.
-            self.logger.debug('P4')
-            self.dma.recvchannel.transfer(self.buff)
-            self.logger.debug('P5')
-            self.dma.recvchannel.wait()
-            self.logger.debug('P6')
-            # End Operation
-            self.tproc_cfg       = 0      
-            
-            if ( (np.max(self.buff - p_mem) )  == 0):
-                self.logger.info('Program Loaded OK')
-            else:
-                self.logger.error('Error Loading Program')
-
-            
-        def getALL(self):
-            print(self.status_axi())
-
-        def status_axi(self):
-            lines = []
-            lines.append('---------------------------------------------')
-            lines.append('--- AXI Registers')
-            for param in ["tproc_ctrl",
-                    "tproc_cfg",
-                    "rand",
-                    "mem_addr",
-                    "mem_len",
-                    "mem_dt_i",
-                    "mem_dt_o",
-                    "port_lsw",
-                    "port_msw",
-                    "tproc_ext_dt1_i",
-                    "tproc_ext_dt2_i",
-                    "tproc_ext_dt1_o",
-                    "tproc_ext_dt2_o",
-                    "time_usr"]:
-                lines.append("%-16s: %d" % (param.upper(), getattr(self, param)))
-            lines.append('TPROC_STATUS : {0:12d} - {0:039_b}'.format(self.tproc_status))
-            lines.append('TPROC_DEBUG  : {0:12d} - {0:039_b}'.format(self.tproc_debug))
-            return "\n".join(lines)
-     
-        def getStatus(self):
-            print(self.status())
-
-        def status(self):
-            lines = []
-            debug_num = self.tproc_debug
-            lines.append('---------------------------------------------')
-            lines.append('--- Debug signals')
-            lines.append('EXT_MEM_ADDR :' + '{:032b}'.format(debug_num)[0:8])
-            lines.append('PMEM_ADDR    :' + '{:032b}'.format(debug_num)[8:16])
-            lines.append('Time Ref     :' + '{:032b}'.format(debug_num)[16:24])
-            lines.append("")
-            lines.append('FIFO[0] Time :' + '{:032b}'.format(debug_num)[24:27])
-            lines[-1] += ('FIFO_OK :'      + '{:032b}'.format(debug_num)[28])
-            lines.append('Header  :'      + '{:032b}'.format(debug_num)[29:32])
-            status_num = self.tproc_status
-            lines.append('---------------------------------------------')
-            lines.append('--- Memory Unit Status signals')
-            lines.append('AXI_Read  :'+ '{:032b}'.format(status_num)[0])
-            lines[-1] += ('AXI_Write :'+ '{:032b}'.format(status_num)[1] )
-            lines.append('ext_P_Mem_EN  :'+ '{:032b}'.format(status_num)[2])
-            lines[-1] += ('ext_P_Mem_WEN :'+ '{:032b}'.format(status_num)[3])
-            lines.append('ext_D_Mem_EN  :'+ '{:032b}'.format(status_num)[4])
-            lines[-1] += ('ext_D_Mem_WEN :'+ '{:032b}'.format(status_num)[5])
-            lines.append('ext_W_Mem_EN  :'+ '{:032b}'.format(status_num)[6])
-            lines[-1] += ('ext_W_Mem_WEN :'+ '{:032b}'.format(status_num)[7])
-            lines.append('--- Processing Unit Status signals')
-            lines.append('FD0_Empty:'+ '{:032b}'.format(status_num)[8])
-            lines[-1] += ('FD1_Empty:'+ '{:032b}'.format(status_num)[10])
-            lines[-1] += ('FD0_Full:'+ '{:032b}'.format(status_num)[9])
-            lines[-1] += ('FD1_Full:'+ '{:032b}'.format(status_num)[11])
-            lines.append('FW0_Empty:'+ '{:032b}'.format(status_num)[12])
-            lines[-1] += ('FW1_Empty:'+ '{:032b}'.format(status_num)[14])
-            lines[-1] += ('FW0_Full:'+ '{:032b}'.format(status_num)[13])
-            lines[-1] += ('FW1_Full:'+ '{:032b}'.format(status_num)[15])
-            lines.append('PMEM_EN:'+ '{:032b}'.format(status_num)[16])
-            lines[-1] += ('DMEM_WE:'+ '{:032b}'.format(status_num)[17])
-            lines[-1] += ('WMEM_WE:'+ '{:032b}'.format(status_num)[18])
-            lines[-1] += ('PORT_WE:'+ '{:032b}'.format(status_num)[19])
-            lines.append('T_en:'+ '{:032b}'.format(status_num)[26])
-            lines[-1] += ('P_en:'+ '{:032b}'.format(status_num)[27])
-            lines[-1] += ('STATE:'+ '{:032b}'.format(status_num)[29:32])
-            return "\n".join(lines)
-
+from .qick_asm import QickConfig, QickProgram
+from .drivers.generator import *
+from .drivers.readout import *
+from .drivers.tproc import *
 
 
 class AxisSwitch(SocIp):
@@ -2300,6 +343,7 @@ class QickSoc(Overlay, QickConfig):
             elif val['driver'] == AxisAvgBuffer:
                 self.avg_bufs.append(getattr(self, key))
 
+        # AxisReadoutV3 isn't a PYNQ-registered IP block, so we add it here
         for buf in self.avg_bufs:
             if buf.readout not in self.readouts:
                 self.readouts.append(buf.readout)
